requirements.txt
prefect==2.19.9
pyyaml>=6.0.1
pydantic<2.0
pysqlite3-binary>=0.5.2.post3
rule-engine>=4.5.0
fastapi>=0.115.0
uvicorn>=0.30.0

